+++ src/pages/settings/Organization/Documents/DocumentsContainer.tsx
@@ -1,11 +1,7 @@
 import React from "react";
-import UserRoleForm from "../../../../components/Forms/UserRole";
 import { Box, Grid, Typography } from "@material-ui/core";
-import { FORM_ACTIONS } from "../../../../models/constants";
 import { FormattedMessage } from "react-intl";
-import InvitedUserTable from "../../../../components/Table/InvitedUser";
 import AddButton from "../../../../components/Dasboard/AddButton";
-import { AUTH_ACTIONS } from "../../../../utils/access/modules/auth/actions";
 import { MODULE_CODES, userHasAccess } from "../../../../utils/access";
 import { ORGANIZATION_ACTIONS } from "../../../../utils/access/modules/organization/actions";
 import DocumentsTable from "../../../../components/Table/DocumentsTable";
